+++ puma/roc.py
@@ -8,7 +8,7 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 from ftag import Label
-from ftag.utils import calculate_rejection_error
+from ftag.utils import calculate_efficiency_error, calculate_rejection_error
 
 from puma.plot_base import PlotBase, PlotLineObject
 from puma.utils import get_good_colours, get_good_linestyles, logger
@@ -77,6 +77,7 @@ class Roc(PlotLineObject):
         signal_class: str | None = None,
         key: str | None = None,
         ratio_group: str | None = None,
+        use_bkg_eff: bool = False,
         **kwargs,
     ) -> None:
         """Initialise properties of roc curve object.
@@ -99,6 +100,9 @@ class Roc(PlotLineObject):
             Identifier for roc curve e.g. tagger, by default None
         ratio_group : str, optional
             Identifies the reference ROC group for ratio calculation, by default None
+        use_bkg_eff : bool
+            Correct error calculation for background efficiency ROCs instead of
+            background rejection, by default False
         **kwargs : kwargs
             Keyword arguments passed to `puma.PlotLineObject`
 
@@ -120,6 +124,7 @@ class Roc(PlotLineObject):
         self.rej_class = rej_class.name if isinstance(rej_class, Label) else rej_class
         self.key = key
         self.ratio_group = ratio_group or str(rej_class)
+        self.use_bkg_eff = use_bkg_eff
         self.kwargs = kwargs
 
         # Check that the rejection class is a string
@@ -154,6 +159,8 @@ class Roc(PlotLineObject):
             n_test = self.n_test
         if n_test is None:
             raise ValueError("No `n_test` provided, cannot calculate binomial error!")
+        if self.use_bkg_eff:
+            return calculate_efficiency_error(self.bkg_rej[self.non_zero_mask], n_test, norm=norm)
         return calculate_rejection_error(self.bkg_rej[self.non_zero_mask], n_test, norm=norm)
 
     def divide(
@@ -249,6 +256,7 @@ class Roc(PlotLineObject):
             "signal_class": self.signal_class,
             "key": self.key,
             "ratio_group": self.ratio_group,
+            "use_bkg_eff": self.use_bkg_eff,
         })
 
         # Optionally also include any extra kwargs stored on instances